Update fetch from Github
